blackboard/libs/kernel_system/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-kernel"
version = "0.1.0"
description = "Architect-Kernel-Worker system built on LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "jsonpatch>=3.0",
    "jsonschema>=4.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["."]
include = ["langgraph_kernel*"]
